Moved WarpingPathResults
