+++ src/main/java/tsml/classifiers/distance_based/elastic_ensemble/fast_elastic_ensemble/assessingNN/LazyAssessNN_LCSS.java
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.assessingNN;
 
-import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.results.WarpingPathResults;
+import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.WarpingPathResults;
 import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.utils.SequenceStatsCache;
 import tsml.elastic_distance_measures.LCSSDistance;
 import weka.core.Instance;
