+++ src/main/java/tsml/elastic_distance_measures/DTW.java
@@ -3,7 +3,7 @@ DTW with early abandon
  */
 package tsml.elastic_distance_measures;
 
-import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.results.WarpingPathResults;
+import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.WarpingPathResults;
 import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.utils.GenericTools;
 import weka.core.Instance;
