+++ src/main/java/tsml/elastic_distance_measures/LCSSDistance.java
@@ -5,7 +5,7 @@
 
 package tsml.elastic_distance_measures;
 
-import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.results.WarpingPathResults;
+import tsml.classifiers.distance_based.elastic_ensemble.fast_elastic_ensemble.WarpingPathResults;
 import weka.core.EuclideanDistance;
 import weka.core.Instance;
 import weka.core.Instances;
